Begin refactoring start/stop/status commands
