+++ lib/cli/cliStates.js
@@ -2,7 +2,6 @@
 const CLI = require('./messages.js');
 const CLICommand = require('./cliCommand.js');
 const { formatValue } = require('./cliTools');
-const tools = require('../tools.js');
 
 /** Command iobroker state ... */
 module.exports = class CLIStates extends CLICommand {
