+++ lib/cli/index.js
@@ -7,5 +7,6 @@ module.exports = {
     command: {
         object: require('./cliObjects.js'),
         state: require('./cliStates.js'),
+        process: require('./cliProcess.js'),
     }
 };
