+++ lib/cli/messages.js
@@ -1,4 +1,5 @@
 'use strict';
+const tools = require('../tools');
 
 const errorMessages = Object.freeze({
     stateNotFound: (
@@ -45,6 +46,7 @@ const successMessages = Object.freeze({
     adapterStarted: (/** @type {string} */ adapter) => `The adapter "${adapter}" was started.`,
     adapterStopped: (/** @type {string} */ adapter) => `The adapter "${adapter}" was stopped.`,
     adapterRestarted: (/** @type {string} */ adapter) => `The adapter "${adapter}" was restarted.`,
+    controllerStatus: (/** @type {boolean} */ isRunning) => `${tools.appName} is ${isRunning ? '' : 'not '}running.`,
 });
 
 // Capsule success messages in console.log
