+++ lib/setup.js
@@ -27,6 +27,7 @@ let yargs;
 
 function initYargs() {
     yargs = require('yargs')
+            // TODO: These commands should have an explanation
             .usage('Commands:\n' +
                 tools.appName + ' setup [--objects <host>] [--states <host>] [custom]\n' +
                 tools.appName + ' start\n' +
@@ -141,6 +142,8 @@ function processCommand(command, args, params, callback) {
 
     /** @type {import('./cli/cliCommand').CLICommandContext} */
     const commandContext = { dbConnect, callback, showHelp };
+    /** @type {import('./cli/cliCommand').CLICommandOptions} */
+    const commandOptions = Object.assign({}, params, commandContext);
 
     switch (command) {
 
@@ -304,23 +307,8 @@ function processCommand(command, args, params, callback) {
 
         case 'status':
         case 'isrun': {
-            dbConnect(params, (objects, states, isOffline) => {
-                const config = JSON.parse(fs.readFileSync(require.resolve(tools.getConfigFileName())).toString());
-                let exitCode = isOffline ? 100 : undefined;
-                if (isOffline) {
-                    console.log(tools.appName + ' is not running');
-                } else {
-                    console.log(tools.appName + ' is running');
-                }
-                console.log('');
-                if (args[0]) {
-                    showConfig(config);
-                } else {
-                    console.log(`Objects type: ${config.objects.type}`);
-                    console.log(`States  type: ${config.states.type}`);
-                }
-                return void callback(exitCode);
-            });
+            const processCommand = new cli.command.process(commandOptions);
+            processCommand.status(args);
             break;
         }
 
@@ -840,18 +828,14 @@ function processCommand(command, args, params, callback) {
 
         case 'o':
         case 'object': {
-            const objectsCommand = new cli.command.object(
-                Object.assign({}, params, commandContext)
-            );
+            const objectsCommand = new cli.command.object(commandOptions);
             objectsCommand.execute(args);
             break;
         }
 
         case 's':
         case 'state': {
-            const statesCommand = new cli.command.state(
-                Object.assign({}, params, commandContext)
-            );
+            const statesCommand = new cli.command.state(commandOptions);
             statesCommand.execute(args);
             break;
         }
@@ -2535,21 +2519,21 @@ function resetDbConnect(callback) {
     }
 }
 
-function showConfig(config, root) {
-    root = root || [];
-    const prefix = root.join('/').toUpperCase();
-    for (const attr in config) {
-        if (!config.hasOwnProperty(attr)) continue;
-        if (attr.match(/comment$/i)) continue;
-        if (typeof config[attr] === 'object') {
-            const nextRoot = JSON.parse(JSON.stringify(root));
-            nextRoot.push(attr);
-            showConfig(config[attr], nextRoot);
-        } else {
-            console.log(`${prefix}${(prefix ? '/' : '') + attr}: ` + config[attr]);
-        }
-    }
-}
+// function showConfig(config, root) {
+//     root = root || [];
+//     const prefix = root.join('/').toUpperCase();
+//     for (const attr in config) {
+//         if (!config.hasOwnProperty(attr)) continue;
+//         if (attr.match(/comment$/i)) continue;
+//         if (typeof config[attr] === 'object') {
+//             const nextRoot = JSON.parse(JSON.stringify(root));
+//             nextRoot.push(attr);
+//             showConfig(config[attr], nextRoot);
+//         } else {
+//             console.log(`${prefix}${(prefix ? '/' : '') + attr}: ` + config[attr]);
+//         }
+//     }
+// }
 
 /**
  * Connects to the DB or tests the connection. The callback has the following signature:
